Set RNG seed before every test
